+++ packages/dock-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/dock-blockchain-modules
 
+## 0.21.0
+
+### Minor Changes
+
+- Move `getVerificationMethod` logic to the `CustomLinkedDataSignature`
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.32.0
+
 ## 0.20.0
 
 ### Minor Changes